ocr-worker/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
requests==2.32.4
numpy==1.26.4
gunicorn==21.2.0
